Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-    float i;
-    scanf("%f",&i);
+    double i;
+    scanf("%lf",&i);
     printf("You entered: %.4f",i);
 }
